create PlayerFinder
