+++ Assets/Scripts/Player/PlayerFinder.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9dba1bdd7b8c6ca42b9bca90eb15618d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
